Rust: convert current crate into workspace.
The current crate works well as an embedded runtime and abstraction layer for the platform. Now
binaries can add it as a library dependency without having to worry about bindgen and the like, and
can include their own asm for low-level setup on top of the symbols set up by the `tiny-rt` lib.